refactor: move Audits component into own folder
